Code cleaned: Refactore slices
