+++ src/main.jsx
@@ -2,7 +2,7 @@ import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
 import "./index.css";
 import { Provider } from "react-redux";
-import store from "./redux/store.js";
+import store from "./redux/store";
 import 'bootstrap/dist/css/bootstrap.min.css';
