+++ src/redux/slices/blogSlice.jsx
@@ -1,21 +1,22 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
-import { baseUrl } from "../config";
+import { baseUrl } from "../../config";
 
 // Async thunk to fetch all blogs from the views endpoint
-export const fetchBlogs = createAsyncThunk(
-  "blogs/fetchBlogs",
-  async () => {
-    const response = await axios.get(`${baseUrl}/jsonapi/views/blog_listings/page_1?include=field_content_sections`);
-    return response.data;
-  }
-);
+export const fetchBlogs = createAsyncThunk("blogs/fetchBlogs", async () => {
+  const response = await axios.get(
+    `${baseUrl}/jsonapi/views/blog_listings/page_1?include=field_content_sections`
+  );
+  return response.data;
+});
 
 // Async thunk to fetch a single blog by ID
 export const fetchSingleBlog = createAsyncThunk(
   "blogs/fetchSingleBlog",
   async (blogId) => {
-    const response = await axios.get(`${baseUrl}/jsonapi/node/blogs/${blogId}?include=field_content_sections,field_author,field_hero_image`);
+    const response = await axios.get(
+      `${baseUrl}/jsonapi/node/blogs/${blogId}?include=field_content_sections,field_author,field_hero_image`
+    );
     return response.data.data;
   }
 );
