+++ src/redux/slices/contentSlice.jsx
@@ -1,6 +1,6 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
-import { baseUrl } from "../config";
+import { baseUrl } from "../../config";
 
 // Async thunk to fetch data dynamically based on the endpoint
 export const fetchContentData = createAsyncThunk(
