+++ src/redux/store.js
@@ -1,15 +1,15 @@
 import { configureStore } from "@reduxjs/toolkit";
-import contentReducer from "./contentSlice";
-import mauticReducer from "./mauticSlice";
-import footerReducer from "./footerSlice";
-import blogReducer from "./blogSlice"; 
+import contentReducer from "./slices/contentSlice";
+import mauticReducer from "./slices/mauticSlice";
+import footerReducer from "./slices/footerSlice";
+import blogReducer from "./slices/blogSlice";
 
 const store = configureStore({
   reducer: {
     content: contentReducer,
     mautic: mauticReducer,
     footer: footerReducer,
-    blogs: blogReducer, 
+    blogs: blogReducer,
   },
 });
